pyproject.toml
[project]
name = "kumo_bot"
version = "1.3.0.0a1"
description = "A random bot for featured webseries discussion."
authors = [
    {name = "Tech. TTGames",email = "50541739+Tech-TTGames@users.noreply.github.com"}
]
license = {text = "EPL-2.0"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.6.3,<3.0.0)",
    "fanficfare (>=4.48.0,<5.0.0)",
]

[project.urls]
repository = "https://github.com/Tech-TTGames/KumoFeaturedBot"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
yapf = "^0.43.0"
pylint = "^3.3.8"

[tool.yapf]
based_on_style = "google"
column_limit = 120

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
